backend/requirements.txt
fastapi
uvicorn
scikit-learn
pandas
requests
google-generativeai
python-dotenv
joblib
numpy
twilio
python-multipart
gunicorn
